Feature: Imeplemented Registry and Shop models and generated EF Core …

+++ src/Domain/Interfaces/IForceElement.cs
@@ -5,7 +5,7 @@ namespace Lilia.Domain.Interfaces
     public interface IForceElement
     {
         string Name { get; }
-        int Total();
+        long Total();
         void Display(int indentLevel);
         void DisplaySimple(int indentLevel);
     }
